refactor(cli): split diff.go under the size guardrail; fail empty post-write headings
Two Codex P2 (PR #220, on a685b01):

- cli/diff.go was 498 lines, over the AGENTS.md "<~400 LOC" guardrail. Split the
  presentation half (label + value rendering: humanizeLabel, formatValue,
  duration/JSON/truncate/escape helpers, jsonTypeName) into the new diff_format.go;
  diff.go keeps parsing, the diff tree-walk, comparison, and ordering. Pure
  same-package reorg — no behavior change. diff.go 344 LOC, diff_format.go 168 LOC.

- scripts/e2e/...write-review-live-e2e.sh marked the post-write structure valid as
  soon as a "Post-Write Review" section existed, so a bare Findings/Collision
  check/Advisory label with nothing under it passed — contradicting the script's own
  "show a label only for a section with content / no empty headings" contract, and
  caught by no signal (postwrite_forbidden_empty_bucket only matches the old fixed
  "none" strings). Now flag a label present with no substantive items AND no "none"
  bucket as a dangling empty heading -> structure invalid -> the gate fails the run.
  A label carrying a "none" bucket stays handled by the existing forbidden-bucket
  signal (its own gate), so that contract is unchanged. Regression test added.
